+++ pkg/smtpd/main_test.go
@@ -15,10 +15,16 @@ var th smtpd.TransactionHandler = func(tr *smtpd.Transaction) {
 }
 
 func TestMain(m *testing.M) {
-	smtpd.SetReply(smtpd.CodeReady, "Service ready")
-	smtpd.SetReply(smtpd.CodeClosing, "Service closing transmission channel")
+	smtpd.SetReply(smtpd.Ready, "Service ready")
+	smtpd.SetReply(smtpd.Closing, "Service closing transmission channel")
+	smtpd.SetReply(smtpd.NotAvailable, "Service not available, closing transmission channel")
+	smtpd.SetReply(smtpd.Extensions, "OK (extended)")
 	srv := smtpd.NewServer("mockmail", "localhost", "1024", &th, nil)
-	go srv.ListenAndServe(make(chan struct{}))
+	go func() {
+		if err := srv.ListenAndServe(make(chan struct{})); err != nil {
+			panic(err)
+		}
+	}()
 	os.Exit(m.Run())
 }