+++ pkg/smtpd/response_test.go
@@ -8,13 +8,18 @@ import (
 )
 
 func TestResponse(t *testing.T) {
-	response := smtpd.Response{Code: smtpd.Code(250), Msg: "OK"}
+	response := smtpd.Response{Code: smtpd.Code(250), Msg: []string{"OK"}}
 	assert.Equal(t, "250 OK", response.String(), "Response is not parsed correctly")
 	assert.Equal(t, false, response.IsError(), "Response code 250 indicates a success")
 	assert.Equal(t, true, response.IsSuccess(), "Response code 250 indicates a success")
 
-	response = smtpd.Response{Code: smtpd.Code(500), Msg: "Not OK"}
+	response = smtpd.Response{Code: smtpd.Code(500), Msg: []string{"Not OK"}}
 	assert.Equal(t, "500 Not OK", response.String(), "Response is not parsed correctly")
+	assert.Equal(t, true, response.IsError(), "Response code 500 indicates a failure")
+	assert.Equal(t, false, response.IsSuccess(), "Response code 500 indicates a failure")
+
+	response = smtpd.Response{Code: smtpd.Code(500), Msg: []string{"Multi", "Lines"}}
+	assert.Equal(t, "500-Multi\r\n500 Lines", response.String(), "Response is not parsed correctly")
 	assert.Equal(t, true, response.IsError(), "Response code 500 indicates a failure")
 	assert.Equal(t, false, response.IsSuccess(), "Response code 500 indicates a failure")
 }